+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Change Log
 
+* Release 1.2.next in progress
+  * In builders, for overloaded setters (for which we have matching properties), try to pick the "obviously better" match (using `instance?` against the property value), and only throw if we can't pick one [JDATA-25](https://clojure.atlassian.net/browse/JDATA-25).
+  * In builders, ignore setters for which we do not have matching properties [JDATA-24](https://clojure.atlassian.net/browse/JDATA-24).
+
 * Release 1.2.107 on 2024-02-19
   * Update parent pom and tools.logging versions